+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added network_type to PysuiConfiguration profile declarations
+- BETA implementation of Gasless Stablecoin Transactions
 
 ### Fixed